// types/homepage.ts
// import type { Dispatch, SetStateAction } from "react";

export enum MenuName {
  Home = "home",
  AboutUs = "our-story",
  Careers = "join-us",
  Clients = "our-clients",
  ContactUs = "contact-us",
  ZodiacEvents = "digital-and-event",
  ZodiacLabs = "research-and-development",
  ZodiacSolutions = "interactive-and-multimedia",
  ZodiacStudios = "design-and-animation",
}

export interface SectionProps {
  // setPageMode: Dispatch<SetStateAction<MenuName>>;
}